# BilfenAppPhp.RestAPI/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bilfen API",
        "version": "v1"
    },
    "components": {
        "schemas": {
...
            "Role": {
                "type": "object",
                "required": ["Name", "createdAt", "updatedAt"],
                "properties": {
                    "Id": { "type": "integer", "description": "The auto-generated id of the role" },
                    "Name": { "type": "string", "description": "The role's name" },
                    "createdAt": { "type": "string", "format": "date-time", "description": "The role's created date" },
                    "updatedAt": { "type": "string", "format": "date-time", "description": "The role's updated date" }
                },
                "example": {
                    "Id": 0,
                    "Name": "string",
                    "createdAt": "string",
                    "updatedAt": "string"
                }
            },
...
            "CreateUpdateRole": {
                "type": "object",
                "required": ["Name"],
                "properties": {
                    "Id": { "type": "integer", "description": "The auto-generated id of the role" },
                    "Name": { "type": "string", "description": "The role's name" }
                }
            },
...
        }
    },
    "paths": {
...
        "/api/roles/GetRole/{id}": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "parameters": [{
                    "in": "path",
                    "name": "id",
                    "description": "Role Id",
                    "schema": {
                        "type": "integer"
                    },
                    "required": true
                }],
                "operationId": "roleById",
                "responses": {
                    "200": {
                        "description": "Retrives a role by id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/roles/CreateRole": {
            "post": {
                "tags": ["Roles"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUpdateRole"
                            }
                        }
                    }
                },
                "operationId": "createRole",
                "responses": {
                    "200": {
                        "description": "Create a new role data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/roles/EditRole": {
            "put": {
                "tags": ["Roles"],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUpdateRole"
                            }
                        }
                    }
                },
                "operationId": "editRole",
                "responses": {
                    "200": {
                        "description": "Edit an existing role data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Role"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/roles/DeleteRole/{id}": {
            "delete": {
                "tags": ["Roles"],
                "parameters": [{
                    "in": "path",
                    "name": "id",
                    "description": "Role Id",
                    "schema": {
                        "type": "integer"
                    },
                    "required": true
                }],
                "operationId": "deleteRole",
                "responses": {
                    "204": {
                        "description": "Delete an existing role data"
                    }
                }
            }
        },
...
    },
...
}